+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ymohamed <ymohamed@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/06 11:40:32 by ymohamed          #+#    #+#             */
-/*   Updated: 2022/08/06 22:33:04 by ymohamed         ###   ########.fr       */
+/*   Updated: 2022/08/08 11:34:32 by ymohamed         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@
 typedef unsigned long	t_ul;
 int		ft_printf(const char *s, ...);
 void	ft_printhexp(size_t val, char *base, int *n);
-void	ft_printhex(long int val, char *base, int *n);
+void	ft_printhex(unsigned int val, char *base, int *n);
 
 #endif
